+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project aspires to follow [semantic versioning](http://semver.org).
 
 ## 1.1.0 - 2016-03-03
 ### Added
-- Support for mobile responsive express checkout via option on Client
+- Support for mobile responsive express checkout via option and new methods on Client
 - "pry" gem in dev dependencies, so one can `binding.pry`
 
 ## 1.0.0 - 2014-10-31